requirements.txt
appdirs==1.4.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
comtypes==1.1.7
distlib==0.3.0
filelock==3.0.12
idna==2.9
pywin32==227
pywinauto==0.6.8
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
virtualenv==20.0.7
pip~=19.2.3
rsa~=4.0
pyasn1~=0.4.8
pytz~=2020.1
PyYAML~=5.3.1
cachetools~=4.1.0
google-api-python-client~=1.8.3
google-auth-oauthlib
httplib2~=0.17.3
oauthlib~=3.1.0
setuptools~=41.2.0
uritemplate~=3.0.1
html2text~=2020.1.16